+++ src/app/app.ts
@@ -4,21 +4,21 @@ import { ModuleBuilder } from './../modules/builder';
 import { ModuleContainer } from './../modules/container';
 import { registry } from './../registry';
 import { Logger } from './services/logger';
-import { AppEvents, AppOptions, AppModuleConstructor } from './types';
+import { AppEvents, AppOptions, ModuleConstructor } from './types';
 
 export class App {
   private lastReportAt?: number;
 
   private constructor(
     private readonly containers: Map<
-      AppModuleConstructor,
-      ModuleContainer<AppModuleConstructor>
+      ModuleConstructor,
+      ModuleContainer<ModuleConstructor>
     >,
-    private readonly rootModule: AppModuleConstructor,
+    private readonly rootModule: ModuleConstructor,
     private readonly options: Required<AppOptions>,
   ) {}
 
-  static create(rootModule: AppModuleConstructor, options?: AppOptions) {
+  static create(rootModule: ModuleConstructor, options?: AppOptions) {
     const defaultLogger = new Logger();
     const config: Required<AppOptions> = {
       logger: defaultLogger,
@@ -27,7 +27,7 @@ export class App {
 
     const logger = config.logger || defaultLogger;
 
-    const builder = new ModuleBuilder<AppModuleConstructor>(registry);
+    const builder = new ModuleBuilder<ModuleConstructor>(registry);
     const containers = builder.build(rootModule, {
       globalProviders: [{ identifier: IDENTIFIERS.LOGGER, useValue: logger }],
     });