+++ src/registry.ts
@@ -1,7 +1,3 @@
-import { ModuleMeta } from './modules';
-import { AppModuleConstructor } from './app/types';
+import { ModuleConstructor, ModuleMeta } from './app/types';
 
-export const registry = new Map<
-  AppModuleConstructor,
-  ModuleMeta<AppModuleConstructor>
->();
+export const registry = new Map<ModuleConstructor, ModuleMeta>();